pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sfconn"
authors = [
    {name = "Paresh Adhia", email = "padhia+github@gmail.com"},
]
description = "Snowflake connection helper functions"
readme = "README.md"
requires-python = ">=3.11"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "snowflake-connector-python>=3.7.0",
    "pyjwt",
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["pytest"]
snowpark = ["snowflake-snowpark-python"]

[project.urls]
Homepage = "https://github.com/padhia/sfconn"

[tool.setuptools]
packages = ["sfconn"]

[tool.setuptools.dynamic]
version = {attr = "sfconn.__version__"}

[tool.ruff]
line-length = 130
target-version = "py311"
lint.extend-select = ["I"]

[tool.pytest.ini_options]
pythonpath = [ "." ]
